build(amdsmi): regenerate wrapper and update tests for new IP blocks
- Regenerate python and rust wrappers with scripts and update interface
- Add new blocks to kBlockNameMap in tests/amd_smi_test/test_common.cc
- Update static_assert to reference AMDSMI_GPU_BLOCK_UCIE_PCS

+++ projects/amdsmi/tests/amd_smi_test/functional/gpu/ras/err_cnt_read_test.cc
@@ -99,7 +99,7 @@ void TestErrCntRead::Run(void) {
           std::cout << "Block Error Mask: 0x" << std::hex << enabled_mask << std::endl;
         }
       }
-      for (uint32_t b = AMDSMI_GPU_BLOCK_FIRST; b <= AMDSMI_GPU_BLOCK_LAST; b = b * 2) {
+      for (uint64_t b = AMDSMI_GPU_BLOCK_FIRST; b <= AMDSMI_GPU_BLOCK_LAST; b = b * 2) {
         DISPLAY_AMDSMI_API("amdsmi_get_gpu_ecc_status", "gpu=" + std::to_string(i), VERB(STANDARD));
         err = amdsmi_get_gpu_ecc_status(processor_handles_[i], static_cast<amdsmi_gpu_block_t>(b),
                                         &err_state);
